refactor(utils): move envRedaction to src/main/utils
Pure file move plus import updates, no behavior change. Frees the
module for reuse beyond codeCli (next commit uses it from ai/mcp).
